backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = "The backend for the BCIS Application"
authors = ["Parinthapat Pengpun <parinzee@protonmail.com>"]
license = "AGPL-3.0"

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.0.5"
djangorestframework = "^3.13.1"
Markdown = "^3.3.7"
django-filter = "^21.1"
psycopg2-binary = "^2.9.3"
django-colorfield = "^0.7.2"
requests = "^2.28.1"
requests-cache = "^0.9.5"
django-storages = "^1.12.3"
boto3 = "^1.24.35"
sentry-sdk = "^1.8.0"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}

[tool.poetry.dev-dependencies]
black = "^22.3.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
